Add local storage service and link opening feature
This update introduces a local storage service used for saving and retrieving data and includes handling asynchronous operations. Additionally, a new feature to allow in-app opening of URLs is included along with other minor bug fixes and improvements. An external module for notification has also been added to the project.