Refactor service locator and add ResourceKeysDictionary
Service locator constants are updated to use the new ResourceKeysDictionary for managing resources such as host and folder names. A non-used file in the project is removed, and an addition of host and folder name constants was placed in the ResourceKeysDictionary. This change enhances the handling and usage of resources inside the project.Locator

A ToDo comment specifying that a constant value needs to be used was removed in the `ServiceLocator.cs` file as no change was deemed necessary. The existing line of code adheres to the expected standards and does not require modification.