Implement dynamic background loading for OverviewPageView
The code includes a new converter 'AsyncStreamToImageLoader' that allows dynamic loading of images in the 'OverviewPageView'. It also includes modifications to 'BackgroundComponent' which now has an image source property that can be attached to a URL for loading images. This allows the application to change the background dynamically based on the selected profile.