Add feature to remember last selected profile
A new feature has been added to remember the last selected profile. This is done by adding a new constant in the StorageConstants class, which is used to store the name of the last selected profile. Additionally, necessary modifications have been made in the OverviewPageViewModel and ListViewModel classes to accommodate this feature.