Implement hardware ID into system authentication process
A significant enhancement has been introduced to improve the security and uniqueness of the system authentication. The SystemService now fetches the hardware ID (HWID) during the authentication process, which involves the CPU, motherboard, and disk identifiers. The HWID is then used in the Auth method where it is passed along with Login and Password data.