Update to v0.1.0-rc1 (#51)
* #46 Remove redundant command execution in GmlButton

The `Command.Execute` call within the button click event handler was redundant since the command is already executed elsewhere. This change improves code readability and prevents potential double execution of the command.

* Handle disk full errors with new localized messages.

Added a new IsDiskFull resource string in multiple languages to display an error message when the disk is full. Updated the SystemService and OverviewPageViewModel to catch IOException and check for disk full errors using the new IsDiskFull method in SystemService.

* Add error handling for game profile initialization

Introduce a new error message "GameProfileError" in resource files. Implement logic to capture and display this error when initializing game profiles, ensuring errors are properly logged and displayed to the user.

* Add placeholder for Sentry Java logging

Commented-out code added as a placeholder for future Sentry logging implementation. This will facilitate capturing Java log4j-related exceptions once the actual logging code is integrated.

* Update submodule link Gml.Client

* Add GameProfileError constant to ResourceKeysDictionary

This new constant will be used to provide an error message for issues related to game profiles. It ensures that error handling for game profiles is more robust and maintainable.

* Update submodule link Gml.Client

* Solving the error for folders that have UTF-8 characters or spaces in the name

* Fixing warnings and updating the license string in Gml.Launcher.csproj to address the NETSDK1206 warning. Also adding RuntimeIdentifiers for faster builds.

* Add GameProfileError constant to ResourceKeysDictionary

Re-add the GameProfileError constant to ensure error handling for game profiles is defined. This change corrects its previous removal and maintains consistency in the dictionary.

* Log error output from process data receiver

Add a Console.WriteLine statement to log error data from the process. This helps in debugging by providing immediate visibility of any error output that is not filtered by the existing conditions.

* Add SplashScreen view import and update initialization

Importing the SplashScreen view allows the application to display a splash screen during launch. Additionally, updating the OnFrameworkInitializationCompleted method to be asynchronous supports potential future asynchronous operations.

---------

Co-authored-by: dirold2 <mr.did199@mail.ru>